Written tests for twilio SMS alerts
